refactor: auth-user 데코레이터를 commen으로 이동
